♻️ [refactor] APP-884: review fixes for virtual host mapping
Move VirtualHostResourceAccessKind to types/ (plain enum for now,
ExchangeableEnum codegen conversion is left for upstreaming), add
missing null guard on the CoreWebView2 pointer in the native method.
